+++ src/scenes/index.ts
@@ -1,2 +1,1 @@
 export * from './main.scene';
-export * from './mobile-tablet.scene';
